Moved detail/ContainerUtils.hpp to util/ContainerUtils.hpp Added some documentation to the EnsureCapacity function
